#30 Use factory pattern for httpClient
